+++ src/checkers/mod.rs
@@ -40,6 +40,7 @@ pub mod reactjs;
 pub mod tomcat;
 pub mod typo3;
 pub mod wordpress;
+pub mod wp_plugins;
 
 use crate::models::{reqres::UrlResponse, technology::Technology, Finding};
 use log::trace;