+++ src/checkers/wp_plugins/mod.rs
@@ -1,0 +1,5 @@
+//! This module declares all the checkers related to WordPress plugins.
+//! They have nothing special, they are standard checkers, but they are
+//! similar and grouping them in a module helps keeping a clean architecture.
+
+pub mod yoast_seo;